refactor: Extract product interface  into a common file
